@@ -39,9 +39,9 @@
     "babel-jest": "23.6.0",
     "babel-loader": "8.0.4",
     "css-loader": "1.0.1",
-    "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.7.1",
-    "enzyme-to-json": "^3.3.5",
+    "enzyme": "3.8.0",
+    "enzyme-adapter-react-16": "1.7.1",
+    "enzyme-to-json": "3.3.5",
     "express": "4.16.4",
     "helmet": "3.15.0",
     "history": "4.7.2",
@@ -53,10 +53,11 @@
     "nodemon": "1.18.7",
     "npm-run-all": "4.1.5",
     "optimize-css-assets-webpack-plugin": "5.0.1",
-    "react": "16.6.3",
+    "puppeteer": "1.11.0",
+    "react": "16.7.0-alpha.2",
     "react-beautiful-dnd": "10.0.3",
     "react-datepicker": "2.0.0",
-    "react-dom": "16.6.3",
+    "react-dom": "16.7.0-alpha.2",
     "react-dropdown-modal": "1.2.3",
     "react-dropdown-selector": "1.0.0",
     "react-icons": "3.2.2",
